feat: convert homepage to static site
